requirements.txt

Flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
openai==0.28.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
eventlet==0.33.3
anthropic==0.25.1
google-generativeai==0.5.4
aiohttp==3.9.3
openai
